refactor: purchase 패키지를 lottonumber 하위 패키지로 이동
