+++ src/test/java/lotto/lottonumber/purchase/validation/MinCostValidatorTest.java
@@ -1,6 +1,6 @@
-package lotto.purchase.validation;
+package lotto.lottonumber.purchase.validation;
 
-import static lotto.purchase.validation.PurchaseValidator.ERROR_REMAINDER_MESSAGE;
+import static lotto.lottonumber.purchase.validation.PurchaseValidator.ERROR_MIN_COST_MESSAGE;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
@@ -8,16 +8,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class RemainderValidatorTest {
+class MinCostValidatorTest {
 
-    PurchaseValidator validator = new RemainderValidator();
+    PurchaseValidator validator = new MinCostValidator();
 
     @ParameterizedTest
-    @DisplayName("1000원으로 나누어 떨어지면 성공 아니면 EX 발생")
+    @DisplayName("1000원 이상이면 통과 미만이면 EX 발생")
     @CsvSource(value = {"1000:500", "5000:-1"}, delimiter = ':')
-    void remain(String success, String ex) {
+    void minCost(String success, String ex) {
         assertThatNoException().isThrownBy(() -> validator.validate(success));
         assertThatIllegalArgumentException().isThrownBy(() -> validator.validate(ex))
-                .withMessageContaining(ERROR_REMAINDER_MESSAGE);
+                .withMessageContaining(ERROR_MIN_COST_MESSAGE);
     }
 }
